requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.25.0
lpips>=0.1.4
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
einops>=0.7.0
safetensors>=0.4.0
